move product related page title resolvers to product module
