+++ projects/core/src/page/title/resolvers/index.ts
@@ -1,5 +1,1 @@
-export * from './product-page-title.resolver';
-export * from './category-page-title.resolver';
 export * from './content-page-title.resolver';
-export * from './search-page-title.resolver';
-export * from './checkout-page-title.resolver';
