+++ projects/core/src/product/product.module.ts
@@ -7,9 +7,15 @@ import {
 } from './facade/index';
 import { ProductStoreModule } from './store/product-store.module';
 import { ProductOccModule } from './occ/product-occ.module';
+import { pageTitleResolvers } from './services';
 
 @NgModule({
   imports: [ProductOccModule, ProductStoreModule],
-  providers: [ProductService, ProductSearchService, ProductReviewService]
+  providers: [
+    ProductService,
+    ProductSearchService,
+    ProductReviewService,
+    ...pageTitleResolvers
+  ]
 })
 export class ProductModule {}
