+++ projects/core/src/product/services/product-page-title.resolver.ts
@@ -1,10 +1,10 @@
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 import { map, filter, switchMap } from 'rxjs/operators';
-import { RoutingService } from '../../../routing';
-import { ProductService } from '../../../product';
-import { PageType } from '../../../occ';
-import { PageTitleResolver } from './page-title.resolver';
+import { RoutingService } from '../../routing';
+import { ProductService } from '../../product/facade';
+import { PageType } from '../../occ';
+import { PageTitleResolver } from '../../page';
 
 @Injectable({
   providedIn: 'root'
